GameAnalysis/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Client for AI APIs
httpx==0.25.2

# Database
aiosqlite==0.19.0

# Configuration and Environment
python-dotenv==1.0.0

# Data Validation and Serialization
pydantic==2.5.0

# Development and Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Logging (built-in logging module is used, no additional dependencies needed)
